requirements.txt
streamlit
numpy
opencv-python
Pillow
tensorflow
matplotlib
scikit-learn